initial spaces => tabs in mindi (bcornec)
From Wolfgang Rosenauer <wolfgang_at_rosenauer.org>:

For some commands (tar, tr and find) the patch just changes the ordering of
parameters because newer versions (at least on Linux) complain about the
wrong ordering. (For tr this is the blocker on SUSE Linux because the
dependency list cannot be created correctly).

The patch also does some more cleanup (and fixes cleanup as for
mindi.err/) at exit.

Cosmetical changes are some changes to log-messages as (invalid) line
numbers or wrong characters.

Another thing is that image creation now uses the same random name for
the mountpoint as well as for the image and again cleans up in the error
handler. In 1.05 there were mindi mounts active after running mindi.